+++ pfxbrick/__init__.py
@@ -6,7 +6,7 @@ import os
 
 # fmt: off
 __project__ = 'pfxbrick'
-__version__ = '0.8.5'
+__version__ = '0.8.6'
 
 # API version